requirements.txt
streamlit==1.28.0
mysql-connector-python==8.1.0
pandas==2.1.0
bcrypt==4.0.1
plotly==5.17.0
altair==5.2.0
python-dotenv==1.0.0
streamlit-calendar==0.2.0